[Objects] Rename relational data-types to a more meaningful name - inital
