+++ bundles/AdminBundle/Resources/public/js/pimcore/object/tags/manyToManyRelation.js
@@ -11,10 +11,10 @@
  * @license    http://www.pimcore.org/license     GPLv3 and PEL
  */
 
-pimcore.registerNS("pimcore.object.tags.multihref");
-pimcore.object.tags.multihref = Class.create(pimcore.object.tags.abstract, {
+pimcore.registerNS("pimcore.object.tags.manyToManyRelation");
+pimcore.object.tags.manyToManyRelation = Class.create(pimcore.object.tags.abstract, {
 
-    type: "multihref",
+    type: "manyToManyRelation",
     dataChanged: false,
     idProperty: "rowId",
     allowBatchAppend: true,
