+++ bundles/AdminBundle/Resources/views/Admin/Index/index.html.php
@@ -427,7 +427,7 @@ $scripts = array(
     "pimcore/object/classes/data/datetime.js",
     "pimcore/object/classes/data/encryptedField.js",
     "pimcore/object/classes/data/time.js",
-    "pimcore/object/classes/data/href.js",
+    "pimcore/object/classes/data/manyToOneRelation.js",
     "pimcore/object/classes/data/image.js",
     "pimcore/object/classes/data/externalImage.js",
     "pimcore/object/classes/data/hotspotimage.js",
@@ -435,9 +435,9 @@ $scripts = array(
     "pimcore/object/classes/data/video.js",
     "pimcore/object/classes/data/input.js",
     "pimcore/object/classes/data/numeric.js",
-    "pimcore/object/classes/data/objects.js",
-    "pimcore/object/classes/data/multihrefMetadata.js",
-    "pimcore/object/classes/data/objectsMetadata.js",
+    "pimcore/object/classes/data/manyToManyObjectRelation.js",
+    "pimcore/object/classes/data/advancedManyToManyRelation.js",
+    "pimcore/object/classes/data/advancedManyToManyObjectRelation.js",
     "pimcore/object/classes/data/nonownerobjects.js",
     "pimcore/object/classes/data/booleanSelect.js",
     "pimcore/object/classes/data/select.js",
@@ -447,7 +447,7 @@ $scripts = array(
     "pimcore/object/classes/data/checkbox.js",
     "pimcore/object/classes/data/consent.js",
     "pimcore/object/classes/data/slider.js",
-    "pimcore/object/classes/data/multihref.js",
+    "pimcore/object/classes/data/manyToManyRelation.js",
     "pimcore/object/classes/data/table.js",
     "pimcore/object/classes/data/structuredTable.js",
     "pimcore/object/classes/data/country.js",
@@ -543,7 +543,7 @@ $scripts = array(
     "pimcore/object/tags/date.js",
     "pimcore/object/tags/datetime.js",
     "pimcore/object/tags/time.js",
-    "pimcore/object/tags/href.js",
+    "pimcore/object/tags/manyToOneRelation.js",
     "pimcore/object/tags/image.js",
     "pimcore/object/tags/encryptedField.js",
     "pimcore/object/tags/externalImage.js",
@@ -553,11 +553,11 @@ $scripts = array(
     "pimcore/object/tags/input.js",
     "pimcore/object/tags/classificationstore.js",
     "pimcore/object/tags/numeric.js",
-    "pimcore/object/tags/objects.js",
-    "pimcore/object/tags/multihrefMetadata.js",
+    "pimcore/object/tags/manyToManyObjectRelation.js",
+    "pimcore/object/tags/advancedManyToManyRelation.js",
     "pimcore/object/gridcolumn/operator/FieldCollectionGetter.js",
     "pimcore/object/gridcolumn/operator/ObjectBrickGetter.js",
-    "pimcore/object/tags/objectsMetadata.js",
+    "pimcore/object/tags/advancedManyToManyObjectRelation.js",
     "pimcore/object/tags/nonownerobjects.js",
     "pimcore/object/tags/booleanSelect.js",
     "pimcore/object/tags/select.js",
@@ -567,7 +567,7 @@ $scripts = array(
     "pimcore/object/tags/textarea.js",
     "pimcore/object/tags/wysiwyg.js",
     "pimcore/object/tags/slider.js",
-    "pimcore/object/tags/multihref.js",
+    "pimcore/object/tags/manyToManyRelation.js",
     "pimcore/object/tags/table.js",
     "pimcore/object/tags/structuredTable.js",
     "pimcore/object/tags/country.js",
