+++ bundles/CoreBundle/Resources/translations/aliases.json
@@ -131,5 +131,9 @@
   "plugins": "extensions",
   "log_search_message": "message",
   "rgbaColor": "color",
+  "multihref": "many_to_many_relation",
+  "multihrefMetadata": "advanced_many_to_many_relation",
+  "objectsMetadata": "advanced_many_to_many_object_relation",
+  "href": "many_to_one_relation",
   "bool": "boolean"
 }
