+++ bundles/CoreBundle/Resources/translations/en.extended.json
@@ -1,4 +1,9 @@
 {
+  "many_to_one_relation": "Many-To-One Relation",
+  "many_to_many_relation": "Many-To-Many Relation",
+  "many_to_many_object_relation": "Many-To-Many Object Relation",
+  "advanced_many_to_many_relation": "Advanced Many-To-Many Relation",
+  "advanced_many_to_many_object_relation": "Advanced Many-To-Many Object Relation",
   "admin_interface": "Admin Interface",
   "icon_library": "Icon Library",
   "system_performance_stability_warning": "Please do not perform this action unless you are sure what you are doing.<br /><b style='color:red'>This action can have a major impact onto the stability and performance of the entire system, and may causes an overload or complete crash of the server!</b><br /><br />Are you sure?",
@@ -283,7 +288,6 @@
   "variables": "Variables",
   "defaults": "Defaults",
   "wysiwyg": "WYSIWYG",
-  "href": "Href",
   "objects": "Objects",
   "default_controller": "Default controller",
   "default_action": "Default action",
@@ -314,7 +318,6 @@
   "button": "Button",
   "priority": "Priority",
   "end": "End",
-  "multihref": "Multi-Href",
   "select_update": "Select update",
   "stable_updates": "Available stable updates",
   "non_stable_updates": "Available non-stable updates",
@@ -384,7 +387,6 @@
   "objectsMetadata_visible_fields": "Visible Fields",
   "objectsMetadata_type_columnbool": "Column Bool",
   "objectsMetadata_type_multiselect": "Multiselect",
-  "objectsMetadata": "Objects with Metadata",
   "file_explorer_saved_file_error": "Cannot save file",
   "reserved_field_names_error": "Please do not use the following reserved field names: ",
   "use_current_date": "Use current date",
@@ -463,7 +465,6 @@
   "calculatedValue_calculatorclass": "Calculator class",
   "calculatedValue_explanation": "See the official documentation to learn more about which methods need to be implemented",
   "calculatedValue_field": "Calculated Value",
-  "multihrefMetadata": "Multihref Advanced",
   "conversionFactor": "Conversion Factor",
   "conversionOffset": "Conversion Offset",
   "default_unit": "Default Unit",
