+++ doc/Development_Documentation/05_Objects/01_Object_Classes/01_Data_Types/README.md
@@ -44,16 +44,16 @@ The entire list of data types is indicated below:
 | languages                | combo box with multiple select and combo box with multiple select and predefined language list from Zend_Locale                                                                                                                                                                                                                                                                                                                                                                 |
 
 
-### [Relation Datatypes](./09_Relation_Types.md)
+### [Relational Datatypes](./09_Relation_Types.md)
 
 | Name                     | Description                                                                                                                                                                                                                                                                                                                                                                                                                                                                    |
 |--------------------------|---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
-| href                     | reference to a Pimcore document, object or asset                                                                                                                                                                                                                                                                                                                                                                                                                                |
-| objects                  | collection of Pimcore object references                                                                                                                                                                                                                                                                                                                                                                                                                                         |
-| multihrefmedatadata      | collection of references to Pimcore documents, objects, assets with additional metadata on the relation                                                                                                                                                                                                                                                                                                                                                                                                                 |
-| objectsmetadata          | collection of Pimcore object references with additional metadata on the relation                                                                                                                                                                                                                                                                                                                                                                                                                                        |
-| [nonownerobjects](./37_Non_Owner_Objects_Type.md)          | object relations which are owned by a different object                                                                                                                                                                                                                                                                                                                                                                                                                          |
-| multihref                | collection of references to Pimcore documents, objects, assets                                                                                                                                                                                                                                                                                                                                                                                                                  |
+| Many-To-One Relation     | reference to a Pimcore document, object or asset                                                                                                                                                                                                                                                                                                                                                                                                                                |
+| Many-To-Many Relation    | collection of references to Pimcore documents, objects, assets                                                                                                                                                                                                                                                                                                                                                                                                                  |
+| Advanced Many-To-Many Relation | collection of references to Pimcore documents, objects, assets with additional metadata on the relation                                                                                                                                                                                                                                                                                                                                                                                                                 |
+| Many-To-Many Object Relation | collection of Pimcore object references                                                                                                                                                                                                                                                                                                                                                                                                                                         |
+| Advanced Many-To-One Object Relation | collection of Pimcore object references with additional metadata on the relation                                                                                                                                                                                                                                                                                                                                                                                                                                        |
+| [Nonowner Objects](./37_Non_Owner_Objects_Type.md)          | object relations which are owned by a different object                                                                                                                                                                                                                                                                                                                                                                                                                          |
 
 
 ### Structured Datatypes
