+++ models/DataObject/ClassDefinition/Data/Objectbricks.php
@@ -197,8 +197,8 @@ class Objectbricks extends Model\DataObject\ClassDefinition\Data
         // relations but not for objectsMetadata, because they have additional data which cannot be loaded directly from the DB
         if (!$objectFromVersion && method_exists($fielddefinition, 'getLazyLoading')
             && $fielddefinition->getLazyLoading()
-            && !$fielddefinition instanceof DataObject\ClassDefinition\Data\ObjectsMetadata
-            && !$fielddefinition instanceof DataObject\ClassDefinition\Data\MultihrefMetadata
+            && !$fielddefinition instanceof DataObject\ClassDefinition\Data\AdvancedManyToManyObjectRelation
+            && !$fielddefinition instanceof DataObject\ClassDefinition\Data\AdvancedManyToManyRelation
             && !$fielddefinition instanceof DataObject\ClassDefinition\Data\Block) {
 
             //lazy loading data is fetched from DB differently, so that not every relation object is instantiated
